Refactored AnimBP Logic and Aimoffset
I refactored all the aim offset logic so that it would all take place on the Blueprint Thread-Safe, I moved over some of my previously cached Values over to it as well. Then I fetch the Pitch value for the aim offset inside of the linked anim layers so as to avoid an error that pops up when ever we are blending out in which case the MainAnimBP reference will tend to return a null pointer before the blend is complete.