Copy Motion and DAO Aim Offset / Mesh Space Fix
This update fixes an issue with the aimoffset jittering by placing it after the offset root bone. In doing so however, it introduced a problem where the Offset Root Bone was no longer accounting for the Additive Root Offset on the animations applied. I believe the fact that the Additive Root Offset is effected by the Root Offset of something applied to anything but the root could be a bug in the Offset Root Bone. My approach gets around whatever is happening in the Offset Root Bone that would cause this problem.