DAO Overhaul
A new approach was taken for the Layering settings across the board which yield better results, a trade-off was made that makes it feel less dynamic but the overall look of it greatly improves.

A new technique to layering was taken where we apply it after the Offset Root Bone to get fluid aiming, for this to work we needed to manually rotate the Root of the Static Poses to match that of the Post-OffsetRootBone Locomotion.

The Aimoffset had to be re-arranged in order for this new approach to work in the Main DAO Layering System.

The Aimoffset of the Copy Motion had to have it's Aim Offset Calculated and interpolated Post-OffsetRootBone, in other words....after the Offset Root Bone was processed on the Anim Graph and before the Aim Offset for the Yaw was Applied.

A slight alteration was made to compensate for Single-Handed Copy Motion Layering, it was in the previous version but it was re-arranged slightly

Some code that was redundant was removed from the Copy Motion.
Some code still remains that is not actually needed for the Copy Motion Variant.