Per-Character AimOffset/HeldObject Positioning Adjustments
Each Character has a Held Object Data Asset for Offsetting Held Objects in the Hand depending on the weapon. a Socket with the name HeldObject_R will be assumed. No Offset to this Socket is required because the HOD for that Held Object will Offset it in the hand appropriately so that no manual adjustment will be needed when setting up the project.

This also allows you to add extra and differing offsets per-weapon so that there is no need for 50 sockets on your character which is a Win because Sockets have an Exponential Performance Impact.

Each Character also has IK Retargeter Data Assets, LR(LiveRetargeting) that will allow you to offset the position and rotation. When you use this alongside my AimOffset Debugger you will be able to more easily adjust the AimOffset. The Debug Printout is not always accurate since other things related to the character can sometimes interfere with it. For example some characters don't retarget well and will have odd rotational transforms on their hands. I don't know if this is related to how the IK Retargeter works or if it's somehow related to Bone Orientation for the bones on their skeleton.

I'm not 100% sure that these changes won't be broken when setting up a new Project from this Repository.