Minor Performance Optimizations
Reworked Select logic so that things wouldn't be getting processed multiple times when they aren't even being used and reduced the amount of times they are processed overall.

Cached the AnimInstance and replaced function calls for it with a reference instead since we were using it so often in the Blueprints.